requirements.txt
# requirements.txt

altair==5.1.2
attrs==23.1.0
blinker==1.6.3
boto3==1.28.70
botocore==1.31.70
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cmake==3.27.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
ffmpy==0.3.1
filelock==3.12.4
Flask==2.3.2
flush==1.0
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gunicorn==19.10.0
idna==3.4
importlib-metadata==6.8.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==4.24.4
psutil==5.9.6
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.27.2
streamlit-option-menu==0.3.6
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.201
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
zipp==3.17.0
